refactor: add EmptyScheduleTable render and separate schedule table component
